move more stuff into cpp files
