+++ src/database/trigger_queries.cpp
@@ -2,6 +2,8 @@
 #include "database/generated/db_tables.hpp"
 #include "database/generated/db_script.hpp"
 #include "database/db_parsing_utils.hpp"
+#include "scripting/triggers/trigger_types.hpp"
+#include "scripting/triggers/trigger_data.hpp"
 
 #include <fmt/format.h>
 #include <spdlog/spdlog.h>
