+++ src/events/event_publisher.cpp
@@ -10,6 +10,7 @@
 #include <fmt/format.h>
 #include <hiredis.h>
 #include <spdlog/spdlog.h>
+#include <nlohmann/json.hpp>
 
 #include "events/event_types.hpp"
