+++ src/scripting/script_engine.cpp
@@ -19,6 +19,9 @@
 #include "world/time_system.hpp"
 
 #include <fmt/format.h>
+#include <spdlog/spdlog.h>
+#include <sol/error.hpp>
+#include <sol/thread.hpp>
 
 #include <random>
 
@@ -512,6 +515,17 @@ void ScriptEngine::install_timeout_hook(lua_State* L) {
 void ScriptEngine::remove_timeout_hook(lua_State* L) {
     // Remove the debug hook
     lua_sethook(L, nullptr, 0, 0);
+}
+
+ScriptResult<sol::protected_function_result> ScriptEngine::wrap_result(sol::protected_function_result&& result) {
+    if (!result.valid()) {
+        sol::error err = result;
+        last_error_ = err.what();
+        spdlog::error("Lua execution error: {}", last_error_);
+        return std::unexpected(ScriptError::ExecutionFailed);
+    }
+
+    return result;
 }
 
 } // namespace FieryMUD
