+++ src/server/mud_server.hpp
@@ -2,9 +2,6 @@
 
 #include "core/ids.hpp"
 #include "core/result.hpp"
-// #include "../game/loop.hpp" // Game loop functionality integrated into server
-#include "commands/command_system.hpp"
-#include "world/world_manager.hpp"
 
 #include <asio.hpp>
 #include <atomic>
