+++ src/world/world_manager.cpp
@@ -19,6 +19,8 @@
 #include "scripting/script_engine.hpp"
 
 #include <spdlog/spdlog.h>
+#define SOL_ALL_SAFETIES_ON 1
+#include <sol/sol.hpp>
 #include <fstream>
 #include <algorithm>
 #include <queue>
